Refactor to move styling logic to new RouteListItem model
It's better to scope the responsibility to this class
