+++ packages/framework/src/Support/Models/RouteList.php
@@ -38,38 +38,7 @@ class RouteList implements Arrayable
     protected function generate(): array
     {
         return collect(Hyde::routes())->map(function (Route $route): array {
-            return [
-                'page_type' => $this->stylePageType($route->getPageClass()),
-                'source_file' => $this->styleSourcePath($route->getSourcePath(), $route->getPageClass()),
-                'output_file' => $this->styleOutputPath($route->getOutputPath()),
-                'route_key' => $this->styleRouteKey($route->getRouteKey()),
-            ];
+            return (new RouteListItem($route))->toArray();
         })->values()->toArray();
-    }
-
-    protected function stylePageType(string $class): string
-    {
-        return str_starts_with($class, 'Hyde') ? class_basename($class) : $class;
-    }
-
-    /** @param  class-string<\Hyde\Pages\Concerns\HydePage>  $class */
-    protected function styleSourcePath(string $path, string $class): string
-    {
-        return $class::isDiscoverable() ? $path : $this->getDynamicSourceLabel();
-    }
-
-    protected function styleOutputPath(string $path): string
-    {
-        return Site::getOutputDirectory()."/$path";
-    }
-
-    protected function styleRouteKey(string $key): string
-    {
-        return $key;
-    }
-
-    protected function getDynamicSourceLabel(): string
-    {
-        return 'dynamic';
     }
 }
